Convert main.py's cli() to _Client.
